refactor: extract pipeline/report.rs
Move all report types (InstallReport, InstalledItem, PluginResult,
RemoveFilter, RemoveReport, RemovedItem, MissingOutput, StaleHash,
OrphanEntry, OrphanReason, MissingPlugin, SkippedPlugin, DoctorReport,
UpdateMode, UpdateStatus, UpdatedItem, UpdateReport, ListedItem,
ListedBundle, ListReport) and their impls out of the monolithic
pipeline.rs into a focused submodule.

Public API is preserved via 'pub use report::*' in mod.rs so external
imports (main.rs, tests/, sibling modules) keep working unchanged.

Refs #180
